Unify bagl/nbgl shared code
